Move URL validation function to utils package
