+++ packages/cli/src/util/format.ts
@@ -56,17 +56,6 @@ export function assertValidPubkeysHex(pubkeysHex: string[]): void {
   }
 }
 
-export function isValidHttpUrl(urlStr: string): boolean {
-  let url;
-  try {
-    url = new URL(urlStr);
-  } catch (_) {
-    return false;
-  }
-
-  return url.protocol === "http:" || url.protocol === "https:";
-}
-
 /**
  * Parses a file to get a list of bootnodes for a network.
  * Bootnodes file is expected to contain bootnode ENR's concatenated by newlines, or commas for
